move some shared fixtures from test_tutorial to conftest
