+++ sovrin_client/test/cli/test_tutorial.py
@@ -2,18 +2,18 @@ import json
 
 import pytest
 
-from sovrin_client.test import waits
-from stp_core.loop.eventually import eventually
-from plenum.test.cli.helper import exitFromCli, \
-    createAndAssertNewKeyringCreation
 from sovrin_common.exceptions import InvalidConnectionException
 from sovrin_common.constants import ENDPOINT
-from plenum.common.signer_did import DidSigner
 
 from sovrin_client.client.wallet.link import Link, constant
-from sovrin_client.test.cli.helper import getFileLines, prompt_is, doubleBraces, \
-    getTotalConnections, getTotalSchemas, getTotalClaimsRcvd, getTotalAvailableClaims, \
-    newKey, ensureConnectedToTestEnv
+from sovrin_client.test.cli.helper import getFileLines, doubleBraces, \
+    getTotalConnections, getTotalSchemas
+
+from sovrin_client.test.cli.conftest import \
+    acceptInvitation, \
+    connectIfNotAlreadyConnected
+
+from plenum.test.cli.helper import exitFromCli
 
 def getSampleConnectionInvitation():
     return {
@@ -89,118 +89,6 @@ def checkIfInvalidEndpointIsRejected(do, map):
            mapper=map)
 
 
-def agentWithEndpointAdded(be, do, agentCli, agentMap, attrAddedOut):
-    be(agentCli)
-
-    ensureConnectedToTestEnv(be, do, agentCli)
-
-    # TODO these belong in there own test, not
-    # TODO part of standing up agents
-
-    # checkIfInvalidEndpointIsRejected(do, agentMap)
-    # checkIfValidEndpointIsAccepted(do, agentMap, attrAddedOut)
-    do('send ATTRIB dest={remote} raw={endpointAttr}',
-       within=5,
-       expect=attrAddedOut,
-       mapper=agentMap)
-    return agentCli
-
-
-@pytest.fixture(scope="module")
-def faberWithEndpointAdded(be, do, faberCli, faberAddedByPhil,
-                           faberMap, attrAddedOut):
-    agentWithEndpointAdded(be, do, faberCli, faberMap, attrAddedOut)
-
-
-@pytest.fixture(scope="module")
-def acmeWithEndpointAdded(be, do, acmeCli, acmeAddedByPhil,
-                          acmeMap, attrAddedOut):
-    agentWithEndpointAdded(be, do, acmeCli, acmeMap, attrAddedOut)
-
-
-@pytest.fixture(scope="module")
-def thriftWithEndpointAdded(be, do, thriftCli, thriftAddedByPhil,
-                            thriftMap, attrAddedOut):
-    agentWithEndpointAdded(be, do, thriftCli, thriftMap, attrAddedOut)
-
-
-def connectIfNotAlreadyConnected(do, expectMsgs, userCli, userMap):
-    # TODO: Shouldn't this be testing the cli command `status`?
-    if not userCli._isConnectedToAnyEnv():
-        do('connect test', within=3,
-           expect=expectMsgs,
-           mapper=userMap)
-
-
-def setPromptAndKeyring(do, name, newKeyringOut, userMap):
-    do('prompt {}'.format(name), expect=prompt_is(name))
-    do('new wallet {}'.format(name), expect=newKeyringOut, mapper=userMap)
-
-
-@pytest.fixture(scope="module")
-def preRequisite(poolNodesStarted,
-                 faberIsRunning, acmeIsRunning, thriftIsRunning,
-                 faberWithEndpointAdded, acmeWithEndpointAdded,
-                 thriftWithEndpointAdded):
-    pass
-
-
-@pytest.fixture(scope="module")
-def walletCreatedForTestEnv(preRequisite, be, do, earlCLI, connectedToTest):
-    be(earlCLI)
-    createAndAssertNewKeyringCreation(do, "default1")
-    createAndAssertNewKeyringCreation(do, "default2")
-    connectIfNotAlreadyConnected(do, connectedToTest, earlCLI, {})
-    createAndAssertNewKeyringCreation(do, "test2")
-    exitFromCli(do)
-
-
-@pytest.fixture(scope="module")
-def aliceCli(preRequisite, be, do, walletCreatedForTestEnv,
-             aliceCLI, newKeyringOut, aliceMap):
-    be(aliceCLI)
-    setPromptAndKeyring(do, "Alice", newKeyringOut, aliceMap)
-    return aliceCLI
-
-
-@pytest.fixture(scope="module")
-def susanCli(preRequisite, be, do, susanCLI, newKeyringOut, susanMap):
-    be(susanCLI)
-    setPromptAndKeyring(do, "Susan", newKeyringOut, susanMap)
-    return susanCLI
-
-
-@pytest.fixture(scope="module")
-def bobCli(preRequisite, be, do, bobCLI, newKeyringOut, bobMap):
-    be(bobCLI)
-    setPromptAndKeyring(do, "Bob", newKeyringOut, bobMap)
-    return bobCLI
-
-@pytest.fixture(scope="module")
-def faberCli(be, do, faberCLI, newKeyringOut, faberMap):
-    be(faberCLI)
-    setPromptAndKeyring(do, "Faber", newKeyringOut, faberMap)
-    newKey(be, do, faberCLI, seed=faberMap['seed'])
-
-    return faberCLI
-
-@pytest.fixture(scope="module")
-def acmeCli(be, do, acmeCLI, newKeyringOut, acmeMap):
-    be(acmeCLI)
-    setPromptAndKeyring(do, "Acme", newKeyringOut, acmeMap)
-    newKey(be, do, acmeCLI, seed=acmeMap['seed'])
-
-    return acmeCLI
-
-
-@pytest.fixture(scope="module")
-def thriftCli(be, do, thriftCLI, newKeyringOut, thriftMap):
-    be(thriftCLI)
-    setPromptAndKeyring(do, "Thrift", newKeyringOut, thriftMap)
-    newKey(be, do, thriftCLI, seed=thriftMap['seed'])
-
-    return thriftCLI
-
 def testNotConnected(be, do, aliceCli, notConnectedStatus):
     be(aliceCli)
     do('status', expect=notConnectedStatus)
@@ -234,15 +122,6 @@ def testShowFaberInvite(be, do, aliceCli, faberMap):
 def testLoadInviteNotExists(be, do, aliceCli, fileNotExists, faberMap):
     be(aliceCli)
     do('load {invite-not-exists}', expect=fileNotExists, mapper=faberMap)
-
-
-@pytest.fixture(scope="module")
-def faberInviteLoadedByAlice(be, do, aliceCli, loadInviteOut, faberMap):
-    totalConnectionsBefore = getTotalConnections(aliceCli)
-    be(aliceCli)
-    do('load {invite}', expect=loadInviteOut, mapper=faberMap)
-    assert totalConnectionsBefore + 1 == getTotalConnections(aliceCli)
-    return aliceCli
 
 
 def testLoadFaberInvite(faberInviteLoadedByAlice):
@@ -305,20 +184,7 @@ def testAcceptUnSyncedFaberInviteWhenNotConnected(be, do, aliceCli,
 #                       totalSchemas=0, totalClaimsRcvd=0)
 
 
-@pytest.fixture(scope="module")
-def faberInviteSyncedWithoutEndpoint(be, do, aliceCli, faberMap,
-                                     preRequisite,
-                                     faberInviteLoadedByAlice,
-                                     connectedToTest,
-                                     connectionNotYetSynced,
-                                     syncConnectionOutWithoutEndpoint):
-    be(aliceCli)
-    connectIfNotAlreadyConnected(do, connectedToTest, aliceCli, faberMap)
 
-    do('sync {inviter}', within=2,
-       expect=syncConnectionOutWithoutEndpoint,
-       mapper=faberMap)
-    return aliceCli
 
 
 def testSyncFaberInviteWithoutEndpoint(faberInviteSyncedWithoutEndpoint):
@@ -342,22 +208,9 @@ def testShowSyncedFaberInvite(be, do, aliceCli, faberMap, connectionNotYetSynced
        mapper=cp)
 
 
-def syncInvite(be, do, userCli, expectedMsgs, mapping):
-    be(userCli)
-
-    do('sync {inviter}', within=3,
-       expect=expectedMsgs,
-       mapper=mapping)
 
 
-@pytest.fixture(scope="module")
-def faberInviteSyncedWithEndpoint(be, do, faberMap, aliceCLI, preRequisite,
-                                  faberInviteSyncedWithoutEndpoint,
-                                  syncConnectionOutWithEndpoint):
-    cp = faberMap.copy()
-    cp.update(last_synced='<this connection has not yet been synchronized>')
-    syncInvite(be, do, aliceCLI, syncConnectionOutWithEndpoint, cp)
-    return aliceCLI
+
 
 
 def testSyncFaberInvite(faberInviteSyncedWithEndpoint):
@@ -396,35 +249,6 @@ def getSignedRespMsg(msg, signer):
     signature = signer.sign(msg)
     msg["signature"] = signature
     return msg
-
-
-def acceptInvitation(be, do, userCli, agentMap, expect):
-    be(userCli)
-    do("accept request from {inviter}",
-       within=15,
-       mapper=agentMap,
-       expect=expect,
-       not_expect=[
-           "Observer threw an exception",
-           "DID is not yet written to Sovrin"]
-       )
-    li = userCli.agent.wallet.getConnectionBy(nonce=agentMap['nonce'])
-    assert li
-    agentMap['identifier'] = li.localIdentifier
-    agentMap['verkey'] = li.localVerkey
-
-
-@pytest.fixture(scope="module")
-def aliceAcceptedFaberInvitation(be, do, aliceCli, faberMap,
-                                 preRequisite,
-                                 syncedInviteAcceptedWithClaimsOut,
-                                 faberInviteSyncedWithEndpoint):
-    acceptInvitation(be, do, aliceCli, faberMap,
-                 syncedInviteAcceptedWithClaimsOut)
-    do("list connections", within = 10,
-       mapper=faberMap,
-       expect="Faber College")
-    return aliceCli
 
 
 def testAliceAcceptFaberInvitationFirstTime(aliceAcceptedFaberInvitation):
@@ -497,24 +321,7 @@ def testReqClaimNotExists(be, do, aliceCli, faberMap, showClaimNotFoundOut,
        mapper=faberMap)
 
 
-@pytest.fixture(scope="module")
-def aliceRequestedTranscriptClaim(be, do, aliceCli, transcriptClaimMap,
-                                       reqClaimOut, preRequisite,
-                                       aliceAcceptedFaberInvitation):
-    be(aliceCli)
-    totalClaimsRcvdBefore = getTotalClaimsRcvd(aliceCli)
-    do("request claim {name}", within=5,
-       expect=reqClaimOut,
-       mapper=transcriptClaimMap)
 
-    async def assertTotalClaimsRcvdIncreasedByOne():
-        total_claims = len((await aliceCli.agent.prover.wallet.getAllClaimsSignatures()).keys())
-        assert totalClaimsRcvdBefore + 1 == total_claims
-
-    aliceCli.looper.runFor(10)
-    timeout = waits.expectedClaimsReceived()
-    aliceCli.looper.run(
-        eventually(assertTotalClaimsRcvdIncreasedByOne, timeout=timeout))
 
 
 def testAliceReqClaim(aliceRequestedTranscriptClaim):
@@ -743,37 +550,6 @@ def testShowJobApplicationProofReqAfterSetAttr(be, do, aliceCli,
 #              expect=["Nonce not found".format(msg)])
 
 
-def sendProof(be, do, userCli, agentMap, newAvailableClaims, extraMsgs=None):
-    be(userCli)
-
-    expectMsgs = [
-        "Your Proof {proof-req-to-match} "
-        "{claim-ver-req-to-show} was "
-        "received and verified"
-    ]
-    if extraMsgs:
-        expectMsgs.extend(extraMsgs)
-    mapping = {}
-    mapping.update(agentMap)
-    if newAvailableClaims:
-        mapping['new-available-claims'] = newAvailableClaims
-        expectMsgs.append("Available Claim(s): {new-available-claims}")
-
-    do("send proof {proof-req-to-match} to {inviter}",
-       within=7,
-       expect=expectMsgs,
-       mapper=mapping)
-
-
-@pytest.fixture(scope="module")
-def jobApplicationProofSent(be, do, aliceCli, acmeMap,
-                            aliceAcceptedAcmeJobInvitation,
-                            aliceRequestedTranscriptClaim,
-                            aliceSelfAttestsAttributes):
-    totalAvailableClaimsBefore = getTotalAvailableClaims(aliceCli)
-    sendProof(be, do, aliceCli, acmeMap, "Job-Certificate")
-    assert totalAvailableClaimsBefore + 1 == getTotalAvailableClaims(aliceCli)
-
 
 def testAliceSendClaimProofToAcme(jobApplicationProofSent):
     pass
@@ -808,30 +584,7 @@ def testShowJobCertClaim(be, do, aliceCli, jobCertificateClaimMap,
     assert totalSchemasBefore + 1 == getTotalSchemas(aliceCli)
 
 
-@pytest.fixture(scope="module")
-def jobCertClaimRequested(be, do, aliceCli, preRequisite,
-                        jobCertificateClaimMap, reqClaimOut1,
-                        jobApplicationProofSent):
 
-    def removeSchema():
-        inviter = jobCertificateClaimMap["inviter"]
-        connections = aliceCli.activeWallet.getMatchingConnections(inviter)
-        assert len(connections) == 1
-        faberId = connections[0].remoteIdentifier
-        name, version = jobCertificateClaimMap["name"], \
-                        jobCertificateClaimMap["version"]
-        aliceCli.activeWallet._schemas.pop((name, version, faberId))
-
-    # Removing schema to check if it fetches the schema again or not
-    # removeSchema()
-
-    be(aliceCli)
-
-    totalClaimsRcvdBefore = getTotalClaimsRcvd(aliceCli)
-    do("request claim {name}", within=7,
-       expect=reqClaimOut1,
-       mapper=jobCertificateClaimMap)
-    assert totalClaimsRcvdBefore + 1 == getTotalClaimsRcvd(aliceCli)
 
 
 def testReqJobCertClaim(jobCertClaimRequested):
@@ -909,28 +662,8 @@ def testAliceShowProofIncludeSingleClaim(
        within=3)
 
 
-@pytest.fixture(scope="module")
-def bankBasicProofSent(be, do, aliceCli, thriftMap,
-                       aliceAcceptedThriftLoanApplication):
-    mapping = {}
-    mapping.update(thriftMap)
-    mapping["proof-req-to-match"] = "Loan-Application-Basic"
-    extraMsgs = ["Loan eligibility criteria satisfied, "
-                 "please send another claim 'Loan-Application-KYC'"]
-    sendProof(be, do, aliceCli, mapping, None, extraMsgs)
-
-
 def testAliceSendBankBasicClaim(bankBasicProofSent):
     pass
-
-
-@pytest.fixture(scope="module")
-def bankKYCProofSent(be, do, aliceCli, thriftMap,
-                     bankBasicProofSent):
-    mapping = {}
-    mapping.update(thriftMap)
-    mapping["proof-req-to-match"] = "Loan-Application-KYC"
-    sendProof(be, do, aliceCli, mapping, None)
 
 
 def restartCliAndTestWalletRestoration(be, do, cli, connectedToTest):
